+++ custom_components/ovo_energy_au/graphql/queries.py
@@ -215,6 +215,9 @@ GET_ACCOUNT_EXTRAS = """
 query GetAccountExtras($input: GetAccountInfoInput!) {
   GetAccountInfo(input: $input) {
     id
+    flex {
+      hasOnboarded
+    }
     payments {
       amount
       date